add another layer CRTP inheritance hierarchy - make Parser inherit from AbstractParser
